32 bit changed to 32-bit (hyphen in between)
Use a hyphen to link numbers to the noun that comes after them when they constitute the initial part of a compound adjective. The reader will understand that both words work together to modify another noun in this manner. This holds true regardless of whether the number is expressed in words or digits.